+++ packages/api/src/handler.ts
@@ -11,6 +11,7 @@ import {
   handleAdminRooms,
 } from "./admin/adminController";
 import { cronController } from "./cron/cronHandler";
+import { handleJoin, joinDecoder } from "./join/joinController";
 import {
   handleAdminPOST,
   handleHttpErrorResponse,
@@ -19,33 +20,34 @@ import {
   init,
   setup,
   teardown,
-} from "./io/io";
-import { handleJoin, joinDecoder } from "./join/joinController";
+} from "./services/io/io";
 import { handleTwilioWebhook } from "./webhook/webhookController";
 
 setup();
 
-export const join = AWSLambda.wrapHandler(async function join(
+export const join: any = AWSLambda.wrapHandler(async function join(
   event: APIGatewayProxyEvent,
   context: Context,
 ): Promise<APIGatewayProxyResult> {
   return handlePublicPOST<"join">(event, context, handleJoin, joinDecoder);
 });
 
-export const webhookStatus = AWSLambda.wrapHandler(async function webhookStatus(
-  event: APIGatewayProxyEvent,
-  context: Context,
-): Promise<APIGatewayProxyResult> {
-  return handlePublicPOST<"webhook/twilio/status">(
-    event,
-    context,
-    handleTwilioWebhook,
-    JsonDecoder.string,
-    false,
-  );
-});
+export const webhookStatus: any = AWSLambda.wrapHandler(
+  async function webhookStatus(
+    event: APIGatewayProxyEvent,
+    context: Context,
+  ): Promise<APIGatewayProxyResult> {
+    return handlePublicPOST<"webhook/twilio/status">(
+      event,
+      context,
+      handleTwilioWebhook,
+      JsonDecoder.string,
+      false,
+    );
+  },
+);
 
-export const adminRooms = AWSLambda.wrapHandler(async function adminRooms(
+export const adminRooms: any = AWSLambda.wrapHandler(async function adminRooms(
   event: APIGatewayProxyEvent,
   context: Context,
 ): Promise<APIGatewayProxyResult> {
@@ -57,7 +59,7 @@ export const adminRooms = AWSLambda.wrapHandler(async function adminRooms(
   );
 });
 
-export const adminParticipants = AWSLambda.wrapHandler(
+export const adminParticipants: any = AWSLambda.wrapHandler(
   async function adminParticipants(
     event: APIGatewayProxyEvent,
     context: Context,
@@ -71,7 +73,7 @@ export const adminParticipants = AWSLambda.wrapHandler(
   },
 );
 
-export const cron = AWSLambda.wrapHandler(async function cron(
+export const cron: any = AWSLambda.wrapHandler(async function cron(
   e: APIGatewayProxyEvent,
   context: Context,
 ) {
@@ -80,7 +82,7 @@ export const cron = AWSLambda.wrapHandler(async function cron(
     await cronController();
 
     await teardown(context);
-    return handleSuccessResponse(null);
+    return handleSuccessResponse({ payload: null });
   } catch (err) {
     await teardown(context);
     return handleHttpErrorResponse(err, e);