+++ packages/react/src/components/NetworkQualityLevel/useParticipantNetworkQualityLevel.ts
@@ -19,10 +19,11 @@ export default function useParticipantNetworkQualityLevel(
       handleNetworkQualityLevelChange,
     );
     return () => {
-      participant.off(
-        "networkQualityLevelChanged",
-        handleNetworkQualityLevelChange,
-      );
+      // TODO
+      // participant.off(
+      //   "networkQualityLevelChanged",
+      //   handleNetworkQualityLevelChange,
+      // );
     };
   }, [participant]);