+++ packages/react/src/components/Participant/Participant.tsx
@@ -1,3 +1,4 @@
+import type { BlameDictionary } from "@amfa-team/user-service";
 import React from "react";
 import type { IParticipant } from "../../entities/Participant";
 import ParticipantInfo from "../ParticipantInfo/ParticipantInfo";
@@ -9,6 +10,7 @@ interface ParticipantProps {
   isLocalParticipant: boolean;
   hideParticipant: boolean;
   loading: boolean;
+  blameDictionary: BlameDictionary;
 }
 
 export default function Participant({
@@ -17,6 +19,7 @@ export default function Participant({
   isLocalParticipant,
   hideParticipant,
   loading,
+  blameDictionary,
 }: ParticipantProps) {
   return (
     <ParticipantInfo
@@ -25,6 +28,7 @@ export default function Participant({
       hideParticipant={hideParticipant}
       loading={loading}
       isLocalParticipant={isLocalParticipant}
+      blameDictionary={blameDictionary}
     >
       {participant && (
         <ParticipantTracks