+++ packages/react/src/components/ParticipantInfo/Controls/ParticipantControls.tsx
@@ -1,10 +1,11 @@
+import type { BlameDictionary } from "@amfa-team/user-service";
 import { BlameAction } from "@amfa-team/user-service";
 import classnames from "classnames";
 import { motion } from "framer-motion";
 import React, { useEffect, useMemo, useState } from "react";
 import type { IParticipant } from "packages/react/src/entities";
 // import { CamIcon } from "./Icons/CamIcon";
-import { SettingsIcon } from "./Icons/SettingsIcon";
+// import { SettingsIcon } from "./Icons/SettingsIcon";
 // import { SoundIcon } from "./Icons/SoundIcon";
 import classes from "./participantControls.module.css";
 
@@ -12,35 +13,36 @@ interface ParticipantControlsProps {
   participant: IParticipant;
   participants: IParticipant[];
   isLocalParticipant: boolean;
+  blameDictionary: BlameDictionary;
 }
 
-function getAngle(value: number) {
-  return {
-    top: `${(50 - Math.sin(value * Math.PI) * 50).toFixed(2)}%`,
-    left: `${(50 + Math.cos(value * Math.PI) * 50).toFixed(2)}%`,
-  };
-}
+// function getAngle(value: number) {
+//   return {
+//     top: `${(50 - Math.sin(value * Math.PI) * 50).toFixed(2)}%`,
+//     left: `${(50 + Math.cos(value * Math.PI) * 50).toFixed(2)}%`,
+//   };
+// }
 
-function getAnimation(from: number, to: number, steps: number) {
-  const angleInc = (to - from) / steps;
-  const animation: { top: string[]; left: string[] } = {
-    top: [],
-    left: [],
-  };
+// function getAnimation(from: number, to: number, steps: number) {
+//   const angleInc = (to - from) / steps;
+//   const animation: { top: string[]; left: string[] } = {
+//     top: [],
+//     left: [],
+//   };
 
-  for (let i = 0; i <= steps; i += 1) {
-    const angle = getAngle(from + angleInc * i);
-    animation.top.push(angle.top);
-    animation.left.push(angle.left);
-  }
+//   for (let i = 0; i <= steps; i += 1) {
+//     const angle = getAngle(from + angleInc * i);
+//     animation.top.push(angle.top);
+//     animation.left.push(angle.left);
+//   }
 
-  return animation;
-}
+//   return animation;
+// }
 
-const p1 = {
-  opened: getAnimation(5 / 6, 3 / 6, 10),
-  closed: getAnimation(3 / 6, 5 / 6, 10),
-};
+// const p1 = {
+//   opened: getAnimation(5 / 6, 3 / 6, 10),
+//   closed: getAnimation(3 / 6, 5 / 6, 10),
+// };
 
 // const p2 = {
 //   opened: getAnimation(5 / 6, 1 / 6, 10),
@@ -57,12 +59,13 @@ const renderAnimation = {
 };
 
 const renderTransition = { type: "ease", duration: 0.5 };
-const transition = { type: "linear", duration: 0.3, delay: 0 };
+// const transition = { type: "linear", duration: 0.3, delay: 0 };
 
 export function ParticipantControls({
   participant,
   participants,
   isLocalParticipant,
+  blameDictionary,
 }: ParticipantControlsProps): JSX.Element | null {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
 
@@ -82,22 +85,32 @@ export function ParticipantControls({
     [participants, participant],
   );
 
+  // localParticipant is not withing participants
+  const index = participants.findIndex((p) => p.sid === participant.sid) + 1;
+  const isOdd = index % 2 !== 0;
+
   if (isLocalParticipant) {
     return null;
   }
 
   return (
     <motion.div
-      className={classes.root}
+      className={classnames(classes.root)}
       // @ts-ignore
       animate={renderAnimation}
       transition={renderTransition}
     >
       <div
         onClick={toggleMenu}
-        className={classnames(classes.control, classes.toggle)}
+        className={classnames(classes.control, classes.toggle, {
+          [classes.isOdd]: isOdd,
+        })}
       >
-        <SettingsIcon />
+        <BlameAction
+          accusedId={participant.identity}
+          witnesses={witnesses}
+          dictionary={blameDictionary}
+        />
       </div>
       {/* <motion.div
         initial={"closed"}
@@ -117,15 +130,19 @@ export function ParticipantControls({
       >
         <SoundIcon />
       </motion.div> */}
-      <motion.div
+      {/* <motion.div
         initial={"closed"}
         animate={isMenuOpen ? "opened" : "closed"}
         variants={p1}
         transition={transition}
         className={classnames(classes.control)}
       >
-        <BlameAction accusedId={participant.identity} witnesses={witnesses} />
-      </motion.div>
+        <BlameAction
+          accusedId={participant.identity}
+          witnesses={witnesses}
+          dictionary={blameDictionary}
+        />
+      </motion.div> */}
     </motion.div>
   );
 }