+++ packages/react/src/hooks/useTrack.ts
@@ -15,8 +15,9 @@ export default function useTrack(publication: ITrackPublication | undefined) {
       publication.on("subscribed", setTrack);
       publication.on("unsubscribed", removeTrack);
       return () => {
-        publication.off("subscribed", setTrack);
-        publication.off("unsubscribed", removeTrack);
+        // TODO
+        // publication.off("subscribed", setTrack);
+        // publication.off("unsubscribed", removeTrack);
       };
     }